+++ src/settings/BatchGroupConfigurationSettings/BatchVoucherExportsList.js
@@ -27,9 +27,10 @@ export function BatchVoucherExportsList({ batchVoucherExports, format, onNeedMor
         <FolioFormattedTime dateString={d.end || d.start} />
       ),
       // eslint-disable-next-line react/prop-types
-      exportButton: ({ batchVoucherId }) => (
+      exportButton: ({ batchVoucherId, end }) => (
         <ExportVoucherButton
           batchVoucherId={batchVoucherId}
+          fileName={end}
           format={format}
         />
       ),